requirements.txt
fastapi
uvicorn[standard]
google-generativeai
google-genai
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
alembic